Move common code between cli and bundle into internal/launch
