+++ CMakeLists.txt
@@ -75,7 +75,7 @@ target_link_libraries(ego-enclave
 
 add_custom_command(
   OUTPUT ego
-  DEPENDS ego/cmd/ego/main.go ${CMAKE_SOURCE_DIR}/ego/cli/*.go ${CMAKE_SOURCE_DIR}/ego/config/*.go
+  DEPENDS ego/cmd/ego/main.go ${CMAKE_SOURCE_DIR}/ego/cli/*.go ${CMAKE_SOURCE_DIR}/ego/config/*.go ${CMAKE_SOURCE_DIR}/ego/internal/**/*.go
   COMMAND ${CMAKE_SOURCE_DIR}/src/build_ego.sh ${CMAKE_BINARY_DIR} ${PROJECT_VERSION} ${TRIMPATH}
   WORKING_DIRECTORY ${CMAKE_SOURCE_DIR}/ego/cmd/ego)
 add_custom_target(egobuild ALL DEPENDS ego)
